requirements.txt
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.5
Flask==3.1.1
Flask-WTF==1.2.1
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pycparser==2.22
python-dotenv==1.1.1
requests==2.32.4
urllib3==2.5.0
Werkzeug==3.1.3
WTForms==3.1.2
zipp==3.23.0

# Performance testing dependencies
pytest==8.3.4
pytest-benchmark==4.0.0
